+++ src/app/components/Organism/SignupContent/SignupContent.css.ts
@@ -1,0 +1,5 @@
+import { style } from "@vanilla-extract/css";
+
+export const margin = style({
+  marginBottom: "40px",
+});
