+++ src/app/components/Organism/SignupContent/SignupContent.stories.tsx
@@ -1,0 +1,14 @@
+import { Meta, StoryObj } from "@storybook/react";
+import SignupContent from "./SignupContent";
+
+const meta: Meta<typeof SignupContent> = {
+  title: "Organism/SignupForm",
+  component: SignupContent,
+  tags: ["autodocs"],
+};
+
+export default meta;
+
+type Story = StoryObj<typeof SignupContent>;
+
+export const From = {};
